backend/pyproject.toml
[project]
name = "vibe-agent"
version = "1.0.0"
description = "Autonomous Full-Stack Repository Intelligence System"
authors = [{name = "Vibe Coding Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "neo4j>=5.16.0",
    "faiss-cpu>=1.7.4",
    "tree-sitter>=0.20.4",
    "sentence-transformers>=2.3.1",
    "google-generativeai>=0.3.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.3",
    "rich>=13.7.0",
    "numpy>=1.26.3",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.3",
    "mypy>=1.8.0",
    "black>=23.12.1",
    "ruff>=0.1.13",
    "ipython>=8.20.0",
]

gpu = [
    "faiss-gpu>=1.7.4",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
